\name{RWebMA-package}
\alias{RWebMA-package}
\alias{RWebMA}
\docType{package}
\title{
RWebMA
}
\description{
This package is an interface for R of WebMA of Yahoo! Japan. http://developer.yahoo.co.jp/webapi/jlp/ma/v1/parse.html
}
\details{
\tabular{ll}{
Package: \tab RWebMA\cr
Type: \tab Package\cr
Version: \tab 0.0.2\cr
Date: \tab 2010-06-29\cr
License: \tab BSD\cr
LazyLoad: \tab yes\cr
}
}
\author{
Yohei Sato

Maintainer: Yohei Sato <yokkuns@tkul.jp>
}
\references{

}
\keyword{ package }
\seealso{
XML
}
\examples{
#r <- yahooWebMA("<yahoo appid>", "<sentence>")
}
